création HangoutService pour update les statuts sorties modif security.yaml pour renvoie logout vers page login
