+++ config/packages/security.yaml
@@ -24,7 +24,7 @@ security:
             logout:
                 path: app_logout
                 # where to redirect after logout
-                # target: app_any_route
+                target: app_login
             remember_me:
                 secret: '%kernel.secret%'
                 lifetime: 604800
